requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-multipart==0.0.6
pydantic==1.10.11
pdf2image==1.16.3
PyMuPDF
Pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.26.4
python-dateutil==2.8.2
paddlepaddle==2.6.0
paddleocr>=2.7.0.3
